+++ app/__init__.py
@@ -1,3 +1,3 @@
-"""SFX Desk."""
+﻿"""SFX Desk."""
 
-__version__ = "1.2.2"
+__version__ = "1.2.3"